requirements.txt
aiohttp==1.0.5
apturl==0.5.2

discord.py==0.16.12
GitPython==2.1.9
httplib2==0.9.2
PyGithub==1.39
urllib3==1.22